+++ db/migrate/20170629141736_add_ignored_safeties_on_tasks.rb
@@ -1,0 +1,5 @@
+class AddIgnoredSafetiesOnTasks < ActiveRecord::Migration[5.1]
+  def change
+    add_column :tasks, :ignored_safeties, :boolean, default: false, null: false
+  end
+end