+++ app/jobs/shipit/github_sync_job.rb
@@ -16,23 +16,17 @@ module Shipit
 
         @stack.transaction do
           shared_parent.try!(:detach_children!)
-          new_commits.each do |gh_commit|
+          appended_commits = new_commits.map do |gh_commit|
             append_commit(gh_commit)
           end
+          @stack.lock_reverted_commits! if appended_commits.any?(&:revert?)
         end
       end
       CacheDeploySpecJob.perform_later(@stack)
     end
 
     def append_commit(gh_commit)
-      appended_commit = @stack.commits.create_from_github!(gh_commit)
-      if appended_commit.revert?
-        impacted_commits = @stack.undeployed_commits.reverse.drop_while { |c| !appended_commit.revert_of?(c) }
-        impacted_commits.pop # appended_commit
-        impacted_commits.each do |impacted_commit|
-          impacted_commit.update!(locked: true)
-        end
-      end
+      @stack.commits.create_from_github!(gh_commit)
     end
 
     def fetch_missing_commits(&block)